deps/debug_px4/pyproject.toml
[project]
name = "debug-px4"
version = "0.1.0"
description = "PX4 Zenoh debug subscriber with matplotlib visualization"
authors = [{ name = "Hanamy", email = "rongerch@outlook.com" }]
requires-python = ">=3.10"
dependencies = [
  "eclipse-zenoh>=1.0.0",
  "matplotlib",
  "numpy",
]

[project.optional-dependencies]
dev = ["ruff"]

[project.scripts]
debug-px4 = "debug_px4.cli:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E402", "F841", "F401"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
